refactor(api): split Drizzle schema into per-table files
Colocate each table with its relations so the schema can grow
without a single monolithic file, without changing the database.
