+++ apps/api/drizzle/index.ts
@@ -1,5 +1,3 @@
-import * as drizzleRelations from './relations';
 import * as drizzleSchema from './schema';
 
 export const schema = drizzleSchema;
-export const relations = drizzleRelations;
