+++ apps/api/drizzle/schema/index.ts
@@ -1,0 +1,13 @@
+export * from './event.schema';
+export * from './event-attendee.schema';
+export * from './event-wishlist.schema';
+export * from './item.schema';
+export * from './item-taker.schema';
+export * from './secret-santa.schema';
+export * from './secret-santa-user.schema';
+export * from './user.schema';
+export * from './user-email-change-verification.schema';
+export * from './user-email-setting.schema';
+export * from './user-password-verification.schema';
+export * from './user-social.schema';
+export * from './wishlist.schema';
