+++ apps/api/drizzle/schema/user-social.schema.ts
@@ -1,0 +1,34 @@
+import { relations } from 'drizzle-orm';
+import { foreignKey, pgTable, unique, varchar } from 'drizzle-orm/pg-core';
+
+import { timestamps } from '../helpers';
+import { userId, userSocialId } from '../ids';
+import { user } from './user.schema';
+
+export const userSocial = pgTable(
+  'user_social',
+  {
+    id: userSocialId().primaryKey().notNull(),
+    userId: userId('user_id').notNull(),
+    email: varchar({ length: 200 }).notNull(),
+    name: varchar({ length: 200 }),
+    socialId: varchar('social_id', { length: 1000 }).notNull(),
+    socialType: varchar('social_type', { length: 50 }).notNull(),
+    pictureUrl: varchar('picture_url', { length: 1000 }),
+    ...timestamps,
+  },
+  table => [
+    foreignKey({ columns: [table.userId], foreignColumns: [user.id], name: 'user_social_user_id_fkey' }).onDelete(
+      'cascade',
+    ),
+    unique('user_social_user_id_social_type_key').on(table.userId, table.socialType),
+    unique('user_social_social_id_social_type_key').on(table.socialId, table.socialType),
+  ],
+);
+
+export const userSocialRelations = relations(userSocial, ({ one }) => ({
+  user: one(user, {
+    fields: [userSocial.userId],
+    references: [user.id],
+  }),
+}));
