+++ apps/api/src/core/database/database.service.ts
@@ -2,29 +2,27 @@ import type { SQL } from 'bun';
 import type { BunSQLDatabase } from 'drizzle-orm/bun-sql';
 
 import { Inject, Injectable, Logger, type OnModuleDestroy } from '@nestjs/common';
-import { relations as drizzleRelations, schema as drizzleSchema } from '@wishlist/api-drizzle';
+import { schema } from '@wishlist/api-drizzle';
 import { drizzle } from 'drizzle-orm/bun-sql';
 
 import { createSqlClient } from './create-sql-client';
 import { DatabaseConfig } from './database.config';
 import { DATABASE_CONFIG_TOKEN } from './database.module-definitions';
 
-export const mergedSchema = { ...drizzleSchema, ...drizzleRelations };
-
-export type DrizzleDatabase = BunSQLDatabase<typeof mergedSchema>;
+export type DrizzleDatabase = BunSQLDatabase<typeof schema>;
 
 @Injectable()
 export class DatabaseService implements OnModuleDestroy {
   private readonly logger = new Logger(DatabaseService.name);
 
-  public readonly schema: typeof mergedSchema = mergedSchema;
+  public readonly schema: typeof schema = schema;
   public readonly sql: SQL;
   public readonly db: DrizzleDatabase;
 
   constructor(@Inject(DATABASE_CONFIG_TOKEN) public readonly config: DatabaseConfig) {
     this.sql = createSqlClient(this.config);
     this.db = drizzle(this.sql, {
-      schema: mergedSchema,
+      schema,
       casing: 'snake_case',
       logger: this.config.verbose
         ? { logQuery: (query, params) => this.logger.log('SQL Query', { query, params }) }
